requirements.txt
python==3.8
torch==1.10.1
torchvision==0.11.2
# CUDA dependencies are handled by conda and not pip
# conda install cudatoolkit=11.3
# conda install cudnn=8.2
torchsummary==1.5.1
numpy==1.23.2
pandas==1.3.4
matplotlib==3.5.0
scikit-learn==1.0.2
opencv-python==4.5.5.64
pillow==9.0.1
# tkinter is included in standard Python installation